pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redot2koinly"
version = "0.1.0"
description = "Convert Redotpay transaction-history screenshots into a Koinly simple transaction CSV"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pillow",
    "easyocr>=1.7.0",
    "pandas>=1.3.0",
    "python-dateutil",
    "click",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
]

[project.scripts]
redot2koinly = "redot2koinly.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["redot2koinly*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]